create domain: split session to domain logic
